fix: close detail modal
