+++ app/(community)/posts/@detailModal/(.)detail/[postId]/page.tsx
@@ -2,7 +2,7 @@
 
 import { Suspense } from 'react';
 import PostDetailModal from '@domains/community/modal/PostDetailModal';
-import { notFound } from 'next/navigation';
+import { notFound, useRouter } from 'next/navigation';
 
 interface CommunityDetailPageProps {
   params: { postId: string };
@@ -12,14 +12,27 @@ const CommunityDetailPageSlotInterceptor = ({
   params
 }: CommunityDetailPageProps) => {
   const { postId } = params;
+  const { push, back } = useRouter();
 
   if (!postId) {
     notFound();
   }
 
+  const handleModalClose = () => {
+    if (window.history.length > 1) {
+      back();
+    } else {
+      push('/posts');
+    }
+  };
+
   return (
     <Suspense fallback={<div>로딩 중...</div>}>
-      <PostDetailModal postId={postId} isOpen={true} />
+      <PostDetailModal
+        postId={postId}
+        isOpen={true}
+        onClose={handleModalClose}
+      />
     </Suspense>
   );
 };
