+++ app/(community)/posts/detail/[postId]/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Suspense } from 'react';
-import { notFound, useRouter } from 'next/navigation';
+import { notFound } from 'next/navigation';
 import PostDetailModal from '@domains/community/modal/PostDetailModal';
 import CommunityPage from '../../page';
 
@@ -11,26 +11,17 @@ interface PostDetailPageProps {
 
 const PostDetailPage = ({ params }: PostDetailPageProps) => {
   const { postId } = params;
-  const router = useRouter();
 
   if (!postId) {
     notFound();
   }
-
-  const handleModalClose = () => {
-    router.push('/posts');
-  };
 
   return (
     <>
       <CommunityPage />
 
       <Suspense fallback={<div>로딩 중...</div>}>
-        <PostDetailModal
-          postId={postId}
-          isOpen={true}
-          onClose={handleModalClose}
-        />
+        <PostDetailModal postId={postId} isOpen={true} />
       </Suspense>
     </>
   );
